sre-crew/requirements.txt
crewai
langchain_openai
prometheus-api-client
requests
pyyaml
kubernetes
pydantic
